requirements.txt
numpy==1.19.2     
pandas==1.1.5  
scikit-learn==0.24.2
scipy==1.5.4
torch==1.10.0
tqdm==4.64.1 
pyyaml

